Modified the setup and structure for testing. Will beb easier to add test cases in the future, especially for regression
