+++ .github/workflows/gradle.yml
@@ -35,16 +35,22 @@ jobs:
 
       - name: Perform IO redirection test (*NIX)
         if: runner.os == 'Linux'
-        working-directory:  ${{ github.workspace }}/text-ui-test
+        working-directory:  ${{ github.workspace }}/text-ui-test/general_test
         run: ./runtest.sh
 
       - name: Perform IO redirection test (MacOS)
         if: always() && runner.os == 'macOS'
-        working-directory:  ${{ github.workspace }}/text-ui-test
+        working-directory:  ${{ github.workspace }}/text-ui-test/general_test
         run: ./runtest.sh
 
       - name: Perform IO redirection test (Windows)
         if: always() && runner.os == 'Windows'
-        working-directory:  ${{ github.workspace }}/text-ui-test
+        working-directory:  ${{ github.workspace }}/text-ui-test/general_test
         shell: cmd
         run: runtest.bat
+
+      - name: Perform all Regression Tests (Windows)
+        if: always() && runner.os == 'Windows'
+        working-directory: ${{ github.workspace }}/text-ui-test/
+        shell: cmd
+        run: runtestmain.bat
