+++ build.gradle
@@ -17,6 +17,7 @@ dependencies {
 }
 
 test {
+    maxParallelForks 3
     useJUnitPlatform()
 
     testLogging {
