+++ text-ui-test/general_test/ACTUAL.TXT
