+++ text-ui-test/runtestmain.bat
@@ -1,0 +1,11 @@
+@ECHO OFF
+
+REM Navigate to existing tests and call them
+cd general_test
+call runtest.bat
+cd ..
+
+
+pause
+
+
